Fix tarball locations for all
Depending on where the package is originally published, the tarball
url might be one of:

1. hostname/pkg/-/pkg-version.tgz
2. hostname/db/_design/ddoc/_rewrite/pkg/-/pkg-version.tgz

Where 'hostname', 'db', and 'ddoc' may or may not be the ones that are
currently being used.

We always want to translate that into:

1. hostname/pkg/-/pkg-version.tgz if the request was for
   hostname/pkg or hostname/pkg/version, or
2. hostname/db/_design/ddoc/_rewrite/pkg/-/pkg-version.tgz if the
   request was for hostname/db/_design/ddoc/_rewrite/pkg, or
   hostname/db/_design/ddoc/_show/pkg, or
   hostname/db/_design/ddoc/_rewrite/pkg/version
Also, the hostname, db, and ddoc should match the ones requested, since
packages are sometimes published to difference places and then replicated
around to other couches or accessed using other ddocs.