tests: Only use couchdb configs we can read
This is important because in many default couch installs, there are two
config files, the global defaults and an empty local config. And the latter
is often only readable by a couch user. As it's generally empty, we can go
ahead without it, but couch won't do this unless we explicitly tell it to.

So we get a list of configs and verify their readability, and then on the
command line, clear the default list with `-n` and pass in those we could
read along with the config for our fixtures.